Move theme loading code from theme.rs to new theme_load.rs
To make it easier to later introduce a 'plist-load' feature similar
to the existing 'yaml-load' feature, but for themes and not syntaxes.
